@@ -6,7 +6,7 @@ publish_to: "none"
 version: 1.0.24+54
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
   # Specify for inference of dependabot
   flutter: 3.19.6
 
